% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_chromosome_data.R
\docType{data}
\name{gene_chromosome_data}
\alias{gene_chromosome_data}
\title{Data containing chromosome information and the genes located on them.}
\format{
A data frame with 20 rows and 2 columns.
\itemize{
  \item{CHR: The chromosome containing the gene.}
  \item{Gene: The gene located on the chromosome.}
   }
}
\usage{
data(gene_chromosome_data)
}
\description{
A dataset containing the chromosome information of the gene expression and methylation array data to be analysed.
}
\details{
The dataset contains the information of chromosomes 1 AND 2 and the genes located on them.
}
\seealso{
\code{\link{gene_expression_data}}

\code{\link{methylation_data}}
}
\keyword{datasets}
